Fix no /remote and usermod issues
When no datasets are mounted, /remote does not exist, and trying to
chmod it returns with an error. Now we avoid this when there are no
datasets mounted.

On the Emulab cluster, machines are setup with a process running on that
machine as the CloudLab user, which causes usermod to fail because it
cannot move a users home directory who has a running process. So we now
kill all user processes before performing usermod.